backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-cloud-firestore>=2.17.0
google-cloud-aiplatform>=1.50.0
vertexai>=1.50.0
Pillow>=10.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
